Allow "WinDisc" monitor
To keep end-to-end tests working after the 21.0.4 jdk update.

Some AWT classes don't correctly handle "headless" mode. There is no such thing as fully headless for Windows. For services and drivers, code for accessing screen devices is slightly different than for applications that are run by the user.

In JDK-8185862 that difference was handled by disabling the UI for services altogether. But in Android Studio this is how end-to-end tests are run, and run successfully with rendering Studio UI, taking screenshots etc.
Instead of fully reverting JDK-8185862, this change would allow the "WinDisc" monitor to pass the `IsValidMonitor` test.


Bug: 371006595
Bug: JDK-8336862
Change-Id: Ie1b2d096552445b7ca7302a9ab7b16cc2bcb0851